Engine: use C arrays instead of C++ maps for storing materials and textures

+++ Engine/src/Core/Materials/Material.cpp
@@ -10,7 +10,7 @@ namespace Yngin {
 
 	Material::~Material() = default;
 
-	uint32_t Material::getId() const {
+	uint16_t Material::getId() const {
 		return impl->id;
 	}
